Reorganize package's structure
Create dgacollection package
Update Travis CI's config
