+++ .travis.yml
@@ -3,4 +3,5 @@ python:
   - "3.2"
   - "3.3"
   - "3.4"
-script: python test.py
+script:
+  - python -m unittest discover
