+++ dgacollection/DGAMalware.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 from datetime import datetime
 
-from Malware import Malware
+from dgacollection.Malware import Malware
 
 """Super class for all malwares implementing DGAs.
