+++ originals/necurs.py
