+++ originals/ranbyus.c
