+++ originals/symmi.py
